refactor: 심화 과제 리팩토링 완료
