+++ src/components/AlertDialog/index.ts
@@ -1,0 +1,3 @@
+import AlertDialog from './AlertDialog';
+
+export default AlertDialog;
