+++ src/components/CalendarView/index.ts
@@ -1,0 +1,5 @@
+import CalendarView from './CalendarView';
+import MonthView from './MonthView';
+import WeekView from './WeekView';
+
+export { CalendarView, MonthView, WeekView };
